Add AddGradeFragment

+++ app/src/main/java/com/example/teachersassistant/data/model/Grade.kt
@@ -22,7 +22,7 @@ data class Grade(
     @ColumnInfo(name = "student_id")
     val studentId: Long,
     @ColumnInfo(name = "subject_id")
-    val subjectId: Int,
+    val subjectId: Long,
     val grade: Float,
     val note: String
 )
